+++ project-m2/ujo-web-ex/src/main/java/org/ujorm/tools/web/HtmlElement.java
@@ -22,7 +22,7 @@ import java.nio.charset.Charset;
 import java.util.function.Consumer;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.ujorm.tools.web.ao.URequest;
+import org.ujorm.tools.web.context.URequest;
 import org.ujorm.tools.Assert;
 import org.ujorm.tools.Check;
 import org.ujorm.tools.xml.ApiElement;
