+++ project-m2/ujo-web-ex/src/main/java/org/ujorm/tools/web/context/UContext.java
@@ -1,4 +1,4 @@
-package org.ujorm.tools.web.ao;
+package org.ujorm.tools.web.context;
 
 public final class UContext {
