+++ project-m2/ujo-web-ex/src/main/java/org/ujorm/tools/web/json/JsonBuilder.java
@@ -21,7 +21,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.ujorm.tools.web.HtmlElement;
 import org.ujorm.tools.web.ao.ObjectProvider;
-import org.ujorm.tools.web.ao.URequest;
+import org.ujorm.tools.web.context.URequest;
 import org.ujorm.tools.xml.config.HtmlConfig;
 
 /**
