+++ project-m2/ujo-web-ex/src/main/java/org/ujorm/tools/web/report/ReportBuilder.java
@@ -25,7 +25,7 @@ import java.util.stream.Stream;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import org.ujorm.tools.web.ao.URequest;
+import org.ujorm.tools.web.context.URequest;
 import org.ujorm.tools.Assert;
 import org.ujorm.tools.Check;
 import org.ujorm.tools.web.Element;
