+++ project-m2/ujo-web-ex/src/test/java/org/ujorm/tools/web/ao/HttpParameterTest.java
@@ -3,6 +3,8 @@ package org.ujorm.tools.web.ao;
 import java.time.Month;
 import static java.time.Month.JANUARY;
 import org.junit.jupiter.api.Test;
+import org.ujorm.tools.web.context.URequest;
+
 import static org.junit.jupiter.api.Assertions.*;
 import static java.time.Month.*;
